Split up test-arrows and test-mixed-lengths (#21354)
