+++ extensions/amp-base-carousel/0.1/test-e2e/test-mixed-lengths-no-snap.js
@@ -25,7 +25,7 @@ const pageHeight = 600;
 
 describes.endtoend('AMP carousel mixed length slides', {
   testUrl: 'http://localhost:8000/test/manual/amp-base-carousel/' +
-      'mixed-lengths.amp.html',
+      'mixed-lengths-no-snap.amp.html',
   experiments: ['amp-base-carousel', 'layers'],
   initialRect: {width: pageWidth, height: pageHeight},
 }, async env => {
@@ -44,9 +44,9 @@ describes.endtoend('AMP carousel mixed length slides', {
     controller = env.controller;
   });
 
-  // Test mixed lengths with snapping. This is center aligned as that seems
-  // make the most sense for snapping mixed lengths.
-  describe('snap', () => {
+  // Test mixed lengths without snapping. This is start aligned as that seems
+  // make the most sense for non-snapping mixed lengths.
+  describe('no snap', () => {
     const slideOneWidth = 600;
     const slideTwoWidth = 400;
 
@@ -55,27 +55,22 @@ describes.endtoend('AMP carousel mixed length slides', {
 
       // First slide has width 75%, and viewport is 600 pixels wide
       await expect(prop(slides[0], 'offsetWidth')).to.equal(slideOneWidth);
-      await expect(controller.getElementRect(slides[0])).to.include({
-        x: (pageWidth - slideOneWidth) / 2,
-      });
+      await expect(controller.getElementRect(slides[0])).to.include({x: 0});
       await assertSpacerWidth(0, slideOneWidth);
       // Second slide has width 50%, and viewport is 400 pixels wide
       await expect(prop(slides[1], 'offsetWidth')).to.equal(slideTwoWidth);
       await expect(controller.getElementRect(slides[1])).to.include({
-        x: slideOneWidth + (pageWidth - slideOneWidth) / 2,
+        x: slideOneWidth,
       });
       await assertSpacerWidth(1, slideTwoWidth);
     });
 
-    it('should snap on the center point', async() => {
+    it('should scroll freely', async() => {
       const el = await getScrollingElement(controller);
       const slides = await getSlides(controller);
-      const scrollAmount = 1 + (slideOneWidth + slideTwoWidth) / 2;
 
-      await controller.scrollBy(el, {left: scrollAmount});
-      await expect(controller.getElementRect(slides[1])).to.include({
-        x: (pageWidth - slideTwoWidth) / 2,
-      });
+      await controller.scrollBy(el, {left: 10});
+      await expect(controller.getElementRect(slides[0])).to.include({x: -10});
     });
   });
 });
